+++ nvim/lua/user/plugins/neo-tree.lua
@@ -24,9 +24,12 @@ require("neo-tree").setup({
     },
     default_component_configs = {
         icon = {
-            folder_closed = "",
-            folder_open = "",
-            folder_empty = "",
+            folder_closed = "󰉋",
+            folder_open = "󰝰",
+            folder_empty = "󰉖",
+            folder_empty_open = "󰷏",
+            default = "*",
+            highlight = "NeoTreeFileIcon",
         },
         modified = {
             symbol = "",